+++ Domain/Repository/MetadataRepository/DiskMetadataRepository.php
@@ -107,7 +107,10 @@ class DiskMetadataRepository extends InMemoryMetadataRepository
      */
     private function saveToDisk(): PromiseInterface
     {
-        @\unlink($this->file);
+        if (\file_exists($this->file)) {
+            \unlink($this->file);
+        }
+
         \touch($this->file);
         \file_put_contents($this->file, \serialize($this->storedMetadata));
 
@@ -121,7 +124,11 @@ class DiskMetadataRepository extends InMemoryMetadataRepository
      */
     private function loadFromDisk(): PromiseInterface
     {
-        $content = @\file_get_contents($this->file);
+        $content = false;
+        if (\file_exists($this->file)) {
+            $content = \file_get_contents($this->file);
+        }
+
         if (!\is_string($content)) {
             $this->storedMetadata = [];
